Seperate view and model unit tests into their own files
